Refactor pandare.qcows into an internal and public interface to fix warnings about circular imports. Add file hashes and checks on download
